requirements.txt
Django==3.2.16
aiohttp==3.8.3
django-environ==0.9.0
djangorestframework==3.14.0
openpyxl==3.0.10
pandas==1.5.1
psycopg2-binary==2.9.4
pydantic==1.10.2